+++ app/views/simulator/edit.html.erb
@@ -253,7 +253,9 @@
     <div class="layout-body">
       <button class="largeButton btn logixButton btn-xs custom-btn--tertiary" id="resetVerilogCode">Reset Code</button>
       <button class="largeButton btn logixButton btn-xs custom-btn--primary" id="saveVerilogCode">Save Code</button>
-      <div id="verilogOutput"></div>
+      <div id="verilogOutput">
+      This is an experimental module. The code is not saved unless the "Save Code" button is clicked.
+      </div>
     </div>
   </div>
 </div>
